Moving nano::thread_pool in to its own file
