+++ nano/core_test/utility.cpp
@@ -1,6 +1,6 @@
 #include <nano/lib/optional_ptr.hpp>
 #include <nano/lib/rate_limiting.hpp>
-#include <nano/lib/threading.hpp>
+#include <nano/lib/thread_pool.hpp>
 #include <nano/lib/timer.hpp>
 #include <nano/lib/utility.hpp>
 #include <nano/secure/utility.hpp>
