+++ nano/node/signatures.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <nano/lib/threading.hpp>
+#include <nano/lib/thread_pool.hpp>
 #include <nano/lib/utility.hpp>
 
 #include <atomic>
